refactor: 설정 파일 분리
