--- src/main/java/nextstep/WebMvcConfiguration.java
@@ -1,49 +1,0 @@
-package nextstep;
-
-import auth.*;
-import lombok.AllArgsConstructor;
-import nextstep.member.MemberDao;
-import nextstep.member.UserDetailsFactoryImpl;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.web.servlet.config.annotation.InterceptorRegistry;
-import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
-
-import java.util.List;
-
-@Configuration
-@AllArgsConstructor
-public class WebMvcConfiguration implements WebMvcConfigurer {
-    private MemberDao memberDao;
-
-    @Bean
-    public UserDetailsFactory userDetailsFactory() {
-        return new UserDetailsFactoryImpl(memberDao);
-    }
-
-    @Bean
-    public JwtTokenProvider jwtTokenProvider() {
-        return new JwtTokenProvider();
-    }
-
-    @Bean
-    public LoginService loginService() {
-        return new LoginService(jwtTokenProvider(), userDetailsFactory());
-    }
-
-    @Bean
-    public LoginController loginController() {
-        return new LoginController(loginService());
-    }
-
-    @Override
-    public void addInterceptors(InterceptorRegistry registry) {
-        registry.addInterceptor(new AdminInterceptor(jwtTokenProvider()))
-                .addPathPatterns("/admin/**");
-    }
-
-    @Override
-    public void addArgumentResolvers(List argumentResolvers) {
-        argumentResolvers.add(new LoginMemberArgumentResolver(loginService()));
-    }
-}
