+++ src/main/java/nextstep/config/WebMvcConfiguration.java
@@ -1,0 +1,31 @@
+package nextstep.config;
+
+import auth.AdminInterceptor;
+import auth.JwtTokenProvider;
+import auth.LoginMemberArgumentResolver;
+import auth.LoginService;
+import lombok.RequiredArgsConstructor;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.web.servlet.config.annotation.InterceptorRegistry;
+import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
+
+import java.util.List;
+
+@Configuration
+@RequiredArgsConstructor
+public class WebMvcConfiguration implements WebMvcConfigurer {
+
+    private final JwtTokenProvider jwtTokenProvider;
+    private final LoginService loginService;
+
+    @Override
+    public void addInterceptors(InterceptorRegistry registry) {
+        registry.addInterceptor(new AdminInterceptor(jwtTokenProvider))
+                .addPathPatterns("/admin/**");
+    }
+
+    @Override
+    public void addArgumentResolvers(List argumentResolvers) {
+        argumentResolvers.add(new LoginMemberArgumentResolver(loginService));
+    }
+}
